+++ Editor/AssetsTools.cs
@@ -4,26 +4,20 @@ using UnityEngine;
 
 namespace Mane.Editor
 {
-    public static class PrefabReserialiser
+    public static class AssetsTools
     {
-        [MenuItem("Assets/Force Re-serialise Asset(s)", false, 45)]
-        private static void ReserialiseAssets()
-        {
-            AssetDatabase.ForceReserializeAssets(Selection.objects
-                .Select(AssetDatabase.GetAssetPath));
+        [MenuItem("GameObject/Apply Prefab(s) Changes", true, -10)]
+        private static bool ApplyPrefabsChecker() => IsPrefabSelected();
 
-            foreach (Object obj in Selection.objects)
-            {
-                if (obj) EditorUtility.SetDirty(obj);
-            }
-        }
-        
         [MenuItem("GameObject/Apply Prefab(s) Changes", false, -10)]
         private static void ApplyPrefabs()
         {
             foreach (Object obj in Selection.objects)
                 ApplyChanges(obj as GameObject);
         }
+
+        [MenuItem("GameObject/Apply Prefab(s) Transform Changes", true, -9)]
+        private static bool ApplyPrefabsTransformChecker() => IsPrefabSelected();
         
         [MenuItem("GameObject/Apply Prefab(s) Transform Changes", false, -9)]
         private static void ApplyPrefabsTransform()
@@ -32,6 +26,9 @@ namespace Mane.Editor
                 ApplyTransformChanges(obj as GameObject);
         }
 
+        [MenuItem("GameObject/Apply Prefab(s) Changes (+Transform)", true, -8)]
+        private static bool ApplyPrefabsAllChecker() => IsPrefabSelected();
+        
         [MenuItem("GameObject/Apply Prefab(s) Changes (+Transform)", false, -8)]
         private static void ApplyPrefabsAll()
         {
@@ -39,6 +36,33 @@ namespace Mane.Editor
             {
                 ApplyChanges(obj as GameObject);
                 ApplyTransformChanges(obj as GameObject);
+            }
+        }
+        
+        private static bool IsPrefabSelected() => 
+            Selection.objects.Any(PrefabUtility.IsPartOfPrefabInstance);
+        
+        
+        [MenuItem("Assets/Force Re-serialise Asset(s)", false, 45)]
+        private static void SaveAssets()
+        {
+            AssetDatabase.ForceReserializeAssets(Selection.objects
+                .Select(AssetDatabase.GetAssetPath));
+
+            foreach (Object obj in Selection.objects)
+                if (obj) EditorUtility.SetDirty(obj);
+        }
+        
+        [MenuItem("Mane/Force Reserialize All Assets", false, 1100)]
+        private static void ForceSaveAssets()
+        {
+            if (EditorUtility.DisplayDialog(
+                "Force Assets Reserialization",
+                "This may be long operation despite of the size of your project, so be aware! There is no progress bar so it may looks like that Unity is frozen, but be patient, it's working. Proceed operation?",
+                "Yes, go ahead!", "Cancel!"))
+            {
+                AssetDatabase.ForceReserializeAssets();
+                AssetDatabase.Refresh();
             }
         }