+++ Editor/PropertyDrawers/EnumArrayOddsDrawer.cs
@@ -8,10 +8,8 @@ namespace Mane.Inspector.Editor
     [CustomPropertyDrawer(typeof(EnumArrayOddsAttribute))]
     public class EnumArrayOddsDrawer : EnumArrayDrawerBase
     {
-        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
-        {
-            return EditorGUI.GetPropertyHeight(property, label, true);
-        }
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label) => 
+            EditorGUI.GetPropertyHeight(property, label, true);
 
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {