+++ Editor/PropertyDrawers/ReadOnlyPropertyDrawer.cs
@@ -6,10 +6,8 @@ namespace Mane.Inspector.Editor
     [CustomPropertyDrawer(typeof(ReadOnlyAttribute))]
     public class ReadOnlyPropertyDrawer : PropertyDrawer
     {
-        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
-        {
-            return EditorGUI.GetPropertyHeight(property, label, true);
-        }
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label) => 
+            EditorGUI.GetPropertyHeight(property, label, true);
 
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {