+++ Runtime/Attributes/EnumArrayAttribute.cs
@@ -5,11 +5,8 @@ namespace Mane.Inspector
 {
     public class EnumArrayAttribute : PropertyAttribute
     {
-        public Type EnumType { get; private set; }
+        public Type EnumType { get; }
         
-        public EnumArrayAttribute(Type enumType)
-        {
-            EnumType = enumType;
-        }
+        public EnumArrayAttribute(Type enumType) => EnumType = enumType;
     }
 }