+++ Runtime/Attributes/PostfixAttribute.cs
@@ -4,8 +4,8 @@ namespace Mane.Inspector
 {
     public class PostfixAttribute : PropertyAttribute
     {
-        public string Text { get; private set; }
-        public float Width { get; private set; }
+        public string Text { get; }
+        public float Width { get; }
 
         public PostfixAttribute(string text, float width = 0f)
         {