+++ Runtime/Extensions/BaseClassesExtensions.cs
@@ -4,9 +4,6 @@ namespace Mane.Extensions
 {
     public static class BaseClassesExtensions
     {
-        public static Version ToVersion(this string v)
-        {
-            return new Version(v);
-        }
+        public static Version ToVersion(this string v) => new Version(v);
     }
 }